Refactor NewsletterController to separate business logic into service
