+++ backend/src/main/java/org/hackbrooklyn/plaza/dto/NewsletterSubscriptionDTO.java
@@ -1,0 +1,22 @@
+package org.hackbrooklyn.plaza.dto;
+
+import lombok.Data;
+import lombok.RequiredArgsConstructor;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotBlank;
+
+@RequiredArgsConstructor
+@Data
+public class NewsletterSubscriptionDTO {
+
+    @NotBlank
+    private String firstName;
+
+    @NotBlank
+    private String lastName;
+
+    @NotBlank
+    @Email
+    private String email;
+}
