+++ backend/src/main/java/org/hackbrooklyn/plaza/exception/DataConflictException.java
@@ -1,0 +1,19 @@
+package org.hackbrooklyn.plaza.exception;
+
+public class DataConflictException extends RuntimeException {
+
+    public DataConflictException() {
+    }
+
+    public DataConflictException(String message) {
+        super(message);
+    }
+
+    public DataConflictException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public DataConflictException(Throwable cause) {
+        super(cause);
+    }
+}
