--- backend/src/main/java/org/hackbrooklyn/plaza/exception/FoundDataConflictException.java
@@ -1,19 +1,0 @@
-package org.hackbrooklyn.plaza.exception;
-
-public class FoundDataConflictException extends RuntimeException {
-
-    public FoundDataConflictException() {
-    }
-
-    public FoundDataConflictException(String message) {
-        super(message);
-    }
-
-    public FoundDataConflictException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public FoundDataConflictException(Throwable cause) {
-        super(cause);
-    }
-}
