+++ backend/src/main/java/org/hackbrooklyn/plaza/service/NewsletterService.java
@@ -1,0 +1,11 @@
+package org.hackbrooklyn.plaza.service;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import org.hackbrooklyn.plaza.dto.NewsletterSubscriptionDTO;
+
+import java.security.NoSuchAlgorithmException;
+
+public interface NewsletterService {
+
+    void subscribeUser(NewsletterSubscriptionDTO subscriptionData) throws NoSuchAlgorithmException, JsonProcessingException;
+}
